inference/pose.py
"""
CueCatcher Pose Estimation — Production Module

Uses YOLO11n-Pose for person detection + RTMPose-l for 133 whole-body keypoints.
Falls back to MediaPipe if models aren't downloaded yet.

VRAM: ~5 GB total (YOLO ~2GB + RTMPose ~3GB)
Speed: ~60-90 FPS on RTX 3090
"""

import numpy as np
from pathlib import Path
from typing import Optional
from loguru import logger

import torch


class PoseEstimator:
    """
    Two-stage pose estimation:
      1. YOLO11n-Pose → person bounding box + coarse 17 keypoints
      2. RTMPose-l → refined 133 whole-body keypoints (body, hands, face, feet)

    For a single-child scenario, YOLO provides the crop region and
    RTMPose provides the detailed skeleton.
    """

    def __init__(self, model_dir: Path, device: str = "cuda:0"):
        self.model_dir = model_dir
        self.device = device
        self._detector = None      # YOLO11
        self._pose_model = None    # RTMPose
        self._mode = "none"

    def load(self):
        """Load models with fallback chain."""
        # Try YOLO11 + RTMPose
        yolo_path = self.model_dir / "pose" / "yolo11n-pose.pt"
        rtmpose_path = self.model_dir / "pose" / "rtmpose-l.pth"

        if yolo_path.exists():
            try:
                from ultralytics import YOLO
                self._detector = YOLO(str(yolo_path))
                self._detector.to(self.device)
                logger.info(f"  ✅ YOLO11n-Pose loaded from {yolo_path}")
            except Exception as e:
                logger.warning(f"  ⚠️  YOLO11 failed: {e}")

        if rtmpose_path.exists():
            try:
                self._load_rtmpose(rtmpose_path)
                self._mode = "rtmpose"
                logger.info(f"  ✅ RTMPose-l loaded from {rtmpose_path}")
            except Exception as e:
                logger.warning(f"  ⚠️  RTMPose failed: {e}")

        # Fallback to MediaPipe
        if self._mode == "none":
            try:
                import mediapipe as mp
                self._mp_pose = mp.solutions.pose.Pose(
                    static_image_mode=False,
                    model_complexity=2,
                    min_detection_confidence=0.5,
                    min_tracking_confidence=0.5,
                )
                self._mp_hands = mp.solutions.hands.Hands(
                    static_image_mode=False,
                    max_num_hands=2,
                    min_detection_confidence=0.5,
                )
                self._mode = "mediapipe"
                logger.info("  ✅ Pose: MediaPipe fallback")
            except Exception as e:
                logger.error(f"  ❌ No pose model available: {e}")

    def _load_rtmpose(self, model_path: Path):
        """Load RTMPose via MMPose."""
        # MMPose inference requires config + checkpoint
        # For production, use the MMPose Inferencer API
        try:
            from mmpose.apis import MMPoseInferencer
            self._pose_model = MMPoseInferencer(
                pose2d="rtmpose-l_8xb256-420e_body8-384x288",
                pose2d_weights=str(model_path),
                device=self.device,
            )
        except ImportError:
            # Fallback: ONNX inference
            import onnxruntime as ort
            onnx_path = model_path.with_suffix(".onnx")
            if onnx_path.exists():
                self._pose_model = ort.InferenceSession(
                    str(onnx_path),
                    providers=["CUDAExecutionProvider", "CPUExecutionProvider"],
                )
            else:
                raise FileNotFoundError(f"Neither MMPose nor ONNX model found: {model_path}")

    def estimate(self, frame: np.ndarray) -> dict:
        """
        Run pose estimation on a BGR frame.

        Returns:
            {
                "keypoints": np.ndarray (N, 3) — x, y, confidence
                "bbox": np.ndarray (4,) — x1, y1, x2, y2
                "person_score": float
                "num_keypoints": int  — 33 (MediaPipe) or 133 (RTMPose)
            }
        """
        if self._mode == "rtmpose":
            return self._estimate_rtmpose(frame)
        elif self._mode == "mediapipe":
            return self._estimate_mediapipe(frame)
        else:
            return {"keypoints": None, "bbox": None, "person_score": 0.0, "num_keypoints": 0}

    def _estimate_rtmpose(self, frame: np.ndarray) -> dict:
        """Production pose estimation with YOLO + RTMPose."""
        result = {"keypoints": None, "bbox": None, "person_score": 0.0, "num_keypoints": 0}

        # Step 1: Detect person with YOLO
        if self._detector:
            detections = self._detector(frame, verbose=False, conf=0.3)
            if detections and len(detections[0].boxes) > 0:
                # Take the person with highest confidence
                boxes = detections[0].boxes
                best_idx = boxes.conf.argmax()
                bbox = boxes.xyxy[best_idx].cpu().numpy()
                result["bbox"] = bbox
                result["person_score"] = float(boxes.conf[best_idx])

                # If YOLO also gave keypoints, use them as fallback
                if detections[0].keypoints is not None:
                    kps = detections[0].keypoints.data[best_idx].cpu().numpy()
                    result["keypoints"] = kps
                    result["num_keypoints"] = kps.shape[0]

        # Step 2: Run RTMPose for detailed 133-point skeleton
        if self._pose_model and result["bbox"] is not None:
            try:
                if hasattr(self._pose_model, '__call__'):
                    # MMPose Inferencer
                    pose_results = self._pose_model(
                        frame,
                        bboxes=[result["bbox"].tolist()],
                    )
                    if pose_results:
                        pred = next(pose_results)
                        if pred.get("predictions"):
                            kps = pred["predictions"][0][0].get("keypoints", [])
                            scores = pred["predictions"][0][0].get("keypoint_scores", [])
                            if kps:
                                keypoints = np.array([
                                    [x, y, s] for (x, y), s in zip(kps, scores)
                                ])
                                result["keypoints"] = keypoints
                                result["num_keypoints"] = keypoints.shape[0]
            except Exception as e:
                logger.debug(f"RTMPose inference error: {e}")

        return result

    def _estimate_mediapipe(self, frame: np.ndarray) -> dict:
        """Fallback MediaPipe pose estimation."""
        result = {"keypoints": None, "bbox": None, "person_score": 0.0, "num_keypoints": 0}

        rgb = frame[:, :, ::-1].copy()
        mp_result = self._mp_pose.process(rgb)

        if mp_result.pose_landmarks:
            h, w = frame.shape[:2]
            landmarks = mp_result.pose_landmarks.landmark
            keypoints = np.array([
                [lm.x, lm.y, lm.visibility] for lm in landmarks
            ])
            result["keypoints"] = keypoints
            result["num_keypoints"] = keypoints.shape[0]
            result["person_score"] = float(np.mean([lm.visibility for lm in landmarks]))

            # Compute bbox from landmarks
            xs = keypoints[:, 0] * w
            ys = keypoints[:, 1] * h
            result["bbox"] = np.array([xs.min(), ys.min(), xs.max(), ys.max()])

        return result

    def unload(self):
        if hasattr(self, '_mp_pose') and self._mp_pose:
            self._mp_pose.close()
        if hasattr(self, '_mp_hands') and self._mp_hands:
            self._mp_hands.close()
        self._detector = None
        self._pose_model = None
        torch.cuda.empty_cache()
