requirements.txt
# ── Server ──
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# ── ML / Inference ──
torch==2.4.1
torchvision==0.19.1
onnxruntime-gpu==1.19.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# ── Pose Estimation ──
mmengine==0.10.4
mmcv==2.2.0
mmdet==3.3.0
mmpose==1.3.2

# ── Face Analysis ──
insightface==0.7.3

# ── Audio ──
librosa==0.10.2
soundfile==0.12.1
opensmile==2.5.0

# ── Gaze ──
mediapipe==0.10.15

# ── TTS (Voxtral) ──
vllm==0.6.4
transformers==4.46.0
huggingface-hub==0.26.0

# ── Data / Storage ──
redis[hiredis]==5.1.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2

# ── Streaming ──
aiortc==1.9.0
av==12.3.0

# ── Utilities ──
pyyaml==6.0.2
loguru==0.7.2
orjson==3.10.7
scipy==1.14.1
scikit-learn==1.5.2
